first problem race condition fixed
