+++ data/two_test_problems.csv
@@ -32,97 +32,3 @@ Substituting into the second equation:
 d = 130
 
 The distance between City A and City B is 130 miles.",130
-TEST-003,"A sequence is defined recursively by $a_1 = 3$, $a_2 = 7$, and $a_n = a_{n-1} + a_{n-2}$ for $n \geq 3$. What is the remainder when $a_{20}$ is divided by 4?","I'll work out the first several terms of the sequence to look for a pattern.
-
-$a_1 = 3$
-$a_2 = 7$
-$a_3 = a_2 + a_1 = 7 + 3 = 10$
-$a_4 = a_3 + a_2 = 10 + 7 = 17$
-$a_5 = a_4 + a_3 = 17 + 10 = 27$
-$a_6 = a_5 + a_4 = 27 + 17 = 44$
-$a_7 = a_6 + a_5 = 44 + 27 = 71$
-$a_8 = a_7 + a_6 = 71 + 44 = 115$
-
-Let's look at the remainders when divided by 4:
-$a_1 \cong 3 \pmod{4}$
-$a_2 \cong 3 \pmod{4}$
-$a_3 \cong 2 \pmod{4}$
-$a_4 \cong 1 \pmod{4}$
-$a_5 \cong 3 \pmod{4}$
-$a_6 \cong 0 \pmod{4}$
-$a_7 \cong 3 \pmod{4}$
-$a_8 \cong 3 \pmod{4}$
-
-I notice the pattern of remainders repeats with period 6: 3, 3, 2, 1, 3, 0, ...
-
-To find $a_{20} \pmod{4}$, I need to find where 20 falls in this cycle.
-20 ÷ 6 = 3 remainder 2
-So $a_{20}$ is in the same position as $a_2$ in the cycle.
-Therefore, $a_{20} \cong 3 \pmod{4}$
-
-The remainder when $a_{20}$ is divided by 4 is 3.",3
-TEST-004,"In a regular hexagon ABCDEF with side length 1, what is the area of the quadrilateral formed by connecting the midpoints of sides AB, BC, DE, and EF?","Let's establish a coordinate system with the center of the hexagon at the origin, and vertex A at (1, 0).
-
-For a regular hexagon with side length 1, the distance from the center to any vertex is also 1.
-
-The vertices of the hexagon are:
-A = (1, 0)
-B = (0.5, 0.866)
-C = (-0.5, 0.866)
-D = (-1, 0)
-E = (-0.5, -0.866)
-F = (0.5, -0.866)
-
-The midpoints of the sides are:
-AB midpoint = (0.75, 0.433)
-BC midpoint = (0, 0.866)
-DE midpoint = (-0.75, -0.433)
-EF midpoint = (0, -0.866)
-
-The quadrilateral formed by these midpoints has vertices (0.75, 0.433), (0, 0.866), (-0.75, -0.433), and (0, -0.866).
-
-I can calculate the area using the Shoelace formula:
-Area = 1/2 × |x₁(y₂-y₄) + x₂(y₃-y₁) + x₃(y₄-y₂) + x₄(y₁-y₃)|
-
-Substituting the coordinates:
-Area = 1/2 × |0.75(0.866-(-0.866)) + 0((-0.433)-0.433) + (-0.75)((-0.866)-0.866) + 0(0.433-(-0.433))|
-     = 1/2 × |0.75(1.732) + 0 + (-0.75)((-1.732)) + 0|
-     = 1/2 × |1.299 + 1.299|
-     = 1/2 × 2.598
-     = 1.299
-
-The area of the quadrilateral is 1.299, which equals $\frac{3\sqrt{3}}{4}$ or approximately 1.299.",1.299
-TEST-005,"A fair six-sided die is rolled 4 times. What is the probability that the product of the 4 numbers rolled is a multiple of 8?","For a number to be a multiple of 8, it must be divisible by 2³. I need to determine when the product of 4 die rolls is divisible by 8.
-
-A die has numbers 1, 2, 3, 4, 5, and 6. Let's analyze which of these contribute factors of 2:
-- 1: provides factor 2⁰
-- 2: provides factor 2¹
-- 3: provides factor 2⁰
-- 4: provides factor 2²
-- 5: provides factor 2⁰
-- 6: provides factor 2¹
-
-To get a product divisible by 2³ (or 8), I need the sum of the exponents of 2 to be at least 3.
-
-Case 1: Roll at least one 4 (contributing 2²)
-Probability = 1 - (probability of rolling no 4s)
-Probability = 1 - (5/6)⁴ = 1 - 625/1296 = 671/1296
-
-Case 2: Roll no 4s, but at least three 2s or 6s (each contributing 2¹)
-Probability = (5/6)⁴ × (probability of at least three 2s or 6s given no 4s)
-
-The probability of rolling a 2 or 6 given no 4s = 2/5
-The probability of rolling a 1, 3, or 5 given no 4s = 3/5
-
-We need at least three 2s or 6s, so:
-Probability = (5/6)⁴ × [C(4,3) × (2/5)³ × (3/5)¹ + C(4,4) × (2/5)⁴]
-            = 625/1296 × [4 × 8/625 × 3/5 + 16/625]
-            = 625/1296 × [96/625 + 16/625]
-            = 625/1296 × 112/625
-            = 112/1296
-            = 28/324
-            = 7/81
-
-Total probability = 671/1296 + 7/81 = 671/1296 + 112/1296 = 783/1296 = 261/432
-
-The answer in the format of the sum of numerator and denominator is 261 + 432 = 693.",693
